Revert "Class name fix"
